requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0

# Code quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0

# Pre-commit hooks
pre-commit>=3.6.0
